DAO слой разделен по сущностям с соответствующими заменами наименований интерфейсов в сервисе
