+++ src/main/java/org/example/springtask/config/security/UserDetailsServiceImpl.java
@@ -2,7 +2,7 @@ package org.example.springtask.config.security;
 
 import lombok.RequiredArgsConstructor;
 import org.example.springtask.entity.Worker;
-import org.example.springtask.repository.ProjectDAO;
+import org.example.springtask.repository.interfaces.WorkerDAO;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -13,12 +13,12 @@ import org.springframework.transaction.annotation.Transactional;
 @Service("userDetailsServiceImpl")
 public class UserDetailsServiceImpl implements UserDetailsService {
 
-    private final ProjectDAO projectDao;
+    private final WorkerDAO workerDAO;
 
     @Override
     @Transactional(readOnly = true)
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Worker worker = projectDao.getWorkerByEmail(username);
+        Worker worker = workerDAO.getWorkerByEmail(username);
         return SecurityUser.fromUser(worker);
     }
 }
