+++ src/main/java/org/example/springtask/repository/FileRepository.java
@@ -105,28 +105,3 @@ public class FileRepository {
         return stringBuilder.toString();
     }
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
